requirements.txt
boto3-1.34.74 
botocore-1.34.74 
s3transfer-0.10.1
Flask==3.0.2
Werkzeug==3.0.0